requirements.txt
chainlit
langchain
langchain-openai
langchain-community
sentence-transformers
faiss-cpu
python-dotenv
openai
